Check tempfile variable before closing
Agent::Util#unpack_blob might fail to create tempfile and enter
ensure clause when file reference is nil. nil#close error would
mask the underlying error in that case. Added nil-check to fix it.

+++ agent/lib/agent/util.rb
@@ -69,8 +69,10 @@ module Bosh::Agent
           logger.info("Failure unpacking blob: #{e.inspect} #{e.backtrace}")
           raise e
         ensure
-          tf.close
-          tf.unlink
+          if tf
+            tf.close
+            tf.unlink
+          end
         end
 
       end
